requirements.txt
SpeechRecognition>=3.14.0
pyttsx3==2.90
PyAudio==0.2.14
numpy>=1.26.0
sounddevice==0.4.7
customtkinter>=5.2.0
pystray>=0.19.0
Pillow>=10.0.0
edge-tts>=7.0.0
pygame>=2.5.0
